Refactor Output Form Composition
